gen_config: convert to float before resample_poly
See https://github.com/scipy/scipy/issues/15620

+++ src/brutefir/gen_config.py
@@ -136,7 +136,7 @@ def resample_brutefir_coeffs(config, coeffs):
             continue
 
         coeff["data"] = signal.resample_poly(
-            coeff["data"], fs, coeff["rate"], window="hanning"
+            coeff["data"].astype('float64'), fs, coeff["rate"], window="hanning"
         )
         coeff["data"] = coeff["data"].astype(TYPE_S16LE)  # TODO
         coeff["rate"] = fs
